OCI: Assert that image hashes are equal across registries
This commit implements a new test whose purpose is to assert that the
hashes for a certain image-arch-tag combination are equal across all
supported registries.

The idea is simple, but the implementation is a bit involved because
of the many corner cases I've had to deal with.

This test is making certain assumptions.  For example, it is
assuming that all supported architectures have been properly published
in all registries; if it notices that this is not the case, then it
errors out.  It is also only testing images from the "ubuntu"
namespace, because that's the only one that is published on both
registries.

Unfortunately, like the test to check if all supported architectures
are available, this test is also currently failing due to the problem
with our s390x 21.04 images.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>